perfflow: import PerfFlowAspect python prototype
Check in the first version of PerfFlowAspect.
PerfFlowAspect is a simple Aspect-Oriented Programming-based
tool that can cast a cross-cutting performance-analysis
concern or "aspect" across a disparate set of components
used to form a modern-day composite science workflow:
 https://en.wikipedia.org/wiki/Aspect-oriented_programming.

PerfFlowAspect will ultimately provide multiple language
support, butt as a starting point, this commit
only adds Python support.

perfflow/aspect.py contains a python decorator called
critical_path. User will use a decorator syntax, e.g.,
@perfflow.aspect.critical_path() to annotate those
functions that are likely to be on the critical path
of the end-to-end performance of the workflow.
The annotated functions become then the Join points
that can be weaved with PerfFlowAspect:
https://en.wikipedia.org/wiki/Join_point.

perfflow/advice_chrome.py implements an action,
known as an advice in AOP, a piece of code
that is invoked whenever each of the
annotated functions is executed.
This particular "advice" logs a performance event data
in the Google Chrome Tracing Format (CTF):
https://docs.google.com/document/d/1CvAClvFfyA5R-PhYUmn5OOQtYMH4h6I0nSsKchNAySU/preview.

This advice offers a basic set of pointcuts to choose
from: https://en.wikipedia.org/wiki/Pointcut with some
well known constructs
- around (wrap the annotated function and emit a duration
          CTF event around the target function)
- before (emit the begin (ph="B") part of duration event before
          executing the annotated function)
- after (emit the end (ph="E") part of duration event after
         the annotated function returns.
It also offers variations of these pointcuts for
asynchronous execution: around_async, before_async,
and after_async.

Introduce advice_dispatch.py which contains
a factory or dispatcher class to make
it easy to integrate future advice type
extension later. As a new advice class will be
added, developers can simply add this class to
the dictionary that this dispatcher class
maintains: i.e., AdviceDispatcher.advice_dict.

The advice classes will be decorated with the
@perfflowaspect decorator.
It has current two advantages: 1)
making it easy to indicate a class is an advice class;
and 2) allowing us to add more base operations and data
that will be commonly needed by future advice class
extensions to this decorating method in the future.
The placeholder decorating function is placed
in aspect_base.py.

Finally, more Aspect configuration (e.g., which
advice to choose etc) support will be added
using TOML configuration  in the future,
in order to provide more control to PerfFlowAspect
users.

More usage information will be added to README.md.