This PR introduces pip installation to perfflowaspect.
Short-term benefits: The PR supports the setup.py accepted way
of providing installation for Python packages. In the short-term,
this PR allows users to install and use PerfFlowAspect without
needing to port the core code to their own repositories.

Long-term benefits: Supporting standard installation allows for
PerfFlowAspect to be published to PyPi, use setuptools to build
wheel releases, and adds a small bit of polish to ease of use.
The pyproject.toml.bak file has been added to make it easy to
port to future install standards.

TOML files will be phasing out setup.py as the de facto way to
install packages. I've included it here as reference/reminder for
future. It currently can be install with the TOML file via poetry.
The current drawback with pyproject.toml is that it does not allow
for pip install in editable mode. For the moment, TOML files can be
used to install via poetry, which allows for easy publishing to PyPi,
an automated vitual environment development mechanism for isolated
testing, and ease of managing dependencies.

commit #1: Renamed perfflow to perfflowaspect.
commit #2: Addition of a setup.py for pip installation.
commit #3: Addition of pyproject.toml.bak (prepare for future).
commit #4: Addition of Python 3.6 support to install files.
commit #5: Updated AUTHORS and package authors.
commit #6: Style tweaks to match black style standards.
commit #7: Addition of Python 3.6 to package descriptors.