parser: add annotation parsing code
Add flex/bison rules for annotation.
Use Flex and Bison to validate
annotataion like
@critical_path(pointcut="around", scope="foo").

Using these tools seems like a bit
overkill, but this will be a win
ultimately as we will extend the spec
for annotation in the future, and it
will be a matter of adding a few
new tokens and rules to extend our
scanner and parser.

Add a shim layer for validation
and parsing (perfflow_parser.[cpp|hpp]).