parser: add flow support
Introduce flow support to analyze RPC-based
execution better.

A flow is a means by which the calling
thread (from within the enclosing method)
annotates a remote execution issuance.

@annotate(flow='in') tells the weaver
that a flow will "start" soon.
If pointcut='around' or 'around_async',
the weaver is expected to add this flow info
at the entering of the annotated function.

@annotate(flow='out') tells the weaver
that the flow finishes.
If pointcut='around' or 'around_async',
the weaver is expected to add this flow
finish event at the entering
of the annotated function.

@annotate(flow='inout') with pointcut='around'
tells the weaver that a new flow starts
and will end in the remote execution. Then,
the remote execution will start a new
flow which will finish at the exit
of the annotated function. If pointcut
is not 'around' types, this should be
just treated as 'in'.

@annotate(flow='outin') with
pointcut='around' types are the opposite
of the above and if this is used
with a pointcut
whose type is not 'around', it is
treated as 'out'.