Use qualifed name in logging performance event data
Transition from name to qualified name when logging performance event data
of the annotated function. The qualified name provides more fine-grained detail
particularly when it comes to annotating nested classes, methods, or functions.
With this change, we can distinguish event data from similarly-named functions
that may be in different classes.