add developer container
Problem: it is hard to build this for use on a native solution
Solution: create a development container. We intiially targeted a VSCode
devcontainer, but ran into considerable issue that we have not figured out
yet! A good fall-back is to use a basic container that can be provided
via an automated build.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>